refactor(proxy): extract OAuth M2M auth into auth.go
Move Auth0 token source and authenticated HTTP client setup out of
client.go so the ITX client constructor stays focused on wiring.
